Use EmbeddingV1Activity and MainActivity names
