+++ packages/firebase_core/example/android/app/src/main/java/io/flutter/plugins/firebasecoreexample/EmbeddingV1Activity.java
@@ -1,0 +1,13 @@
+package io.flutter.plugins.firebasecoreexample;
+
+import android.os.Bundle;
+import io.flutter.app.FlutterActivity;
+import io.flutter.plugins.GeneratedPluginRegistrant;
+
+public class EmbeddingV1Activity extends FlutterActivity {
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    GeneratedPluginRegistrant.registerWith(this);
+  }
+}
